--- src/ewoksorange/gui/concurrency/_Executor.py
@@ -1,14 +1,0 @@
-class CancellableExecutor:
-    def _cancel_future(self, future) -> bool:
-        """Cancel a pending future"""
-        raise NotImplementedError("Base class")
-
-
-class AbortableExecutor:
-    def _abort_future(self, future) -> bool:
-        """Abort a running future"""
-        raise NotImplementedError("Base class")
-
-
-class CancellableAndAbortableExecutor(CancellableExecutor, AbortableExecutor):
-    pass
